修改cleancode问题

+++ msit/components/llm/msit_llm/bc_analyze/synthezier.py
@@ -151,6 +151,10 @@ class Synthesizer(object):
 
         split_command = shlex.split(command)
 
+        if not split_command:
+            logger.error("Command is empty,please try to run it first")
+            raise ValueError()
+
         known_invalid_command = {
             "rm", "mv", "mkfs", "dd",
             "chown", "chmod",
